backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
supabase==2.3.4
google-generativeai==0.3.2
python-multipart==0.0.6
pydantic==2.5.0
httpx==0.25.2
PyMuPDF==1.23.8
openai==1.3.0
numpy==1.24.3 